// storybook/utilities/generate-data.d.ts
export declare function randomNumber(min: number, max: number): number;
export declare const generateDataSet: (dataLength: number) => {
    value: number;
    key: string;
}[];
export declare const generateMultipleSeries: (quantity: number, dataSetLength?: number) => {
    name: string;
    data: {
        value: number;
        key: string;
    }[];
}[];
//# sourceMappingURL=generate-data.d.ts.map