// storybook/components/LineChart/utilities.d.ts
/// <reference types="react" />
import type { LineChartTooltipContentProps } from '@shopify/polaris-viz';
export declare function formatXAxisLabel(value: any): string;
export declare function formatYAxisLabel(value: any): string;
export declare const renderTooltipContent: ({ data, }: {
    data: LineChartTooltipContentProps['data'];
}) => JSX.Element;
export declare function formatHourlyLabel(value: string): string;
//# sourceMappingURL=utilities.d.ts.map